+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `@distributeedlab/w3p` - EVM fallback provider
+
+### Changed
+- `@distributeedlab/w3p` - docs
+
 ## [0.2.0-rc.10] - 2023-05-11
 ### Changed
 - `@distributeedlab/w3p` - types, to be more flexible with external provider proxy constructors